backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic-settings==2.6.1
sqlalchemy==2.0.36
alembic==1.13.3
psycopg[binary]==3.2.3
redis==5.2.0
celery==5.4.0
python-json-logger==2.0.7
pytest==8.3.3
httpx==0.27.2
passlib[argon2]==1.7.4
PyJWT==2.9.0
python-multipart==0.0.12
openpyxl==3.1.5
sentry-sdk==2.17.0
